Fixed a dependency problem for the Debian-based operating systems
The augeas resources used to modify /etc/default/buildslave did not depend on
buildbot::install::slave. This could cause them to create a configuration file
before the buildbot-slave package was even installed. This in turn caused the
buildbot-slave package install to _not_ overwrite that augeas-generated file,
which in turn caused the buildslave start script to fail, because the
augeas-generated file was missing some important bits and pieces. Fixed this by
adding correct dependency information to the augeas resources.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>